refactor: update stock hook명 product purchase로 변경
